requirements.txt
frappe
misaka
PyGithub